refactor(textarea): move Textarea to its folder with theme interface
#443 #234 #461
