+++ src/lib/components/Flowbite/FlowbiteTheme.ts
@@ -25,6 +25,7 @@ import { FlowbiteSelectTheme } from '../Select';
 import { FlowbiteSidebarTheme } from '../Sidebar';
 import { FlowbiteSpinnerTheme } from '../Spinner';
 import { FlowbiteTabTheme } from '../Tab';
+import { FlowbiteTextareaTheme } from '../Textarea';
 import { FlowbiteTextInputTheme } from '../TextInput';
 import { FlowbiteToastTheme } from '../Toast';
 import { FlowbiteTooltipTheme } from '../Tooltip';
@@ -64,6 +65,7 @@ export interface FlowbiteTheme extends Record<string, unknown> {
   radio: FlowbiteRadioTheme;
   select: FlowbiteSelectTheme;
   textInput: FlowbiteTextInputTheme;
+  textarea: FlowbiteTextareaTheme;
 }
 
 export interface FlowbiteBoolean {
