+++ src/lib/components/FormControls/index.ts
@@ -1,3 +1,2 @@
 export * from './HelperText';
-export * from './Textarea';
 export * from './ToggleSwitch';
